pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.3",
    "langchain-google-genai>=3.0.1",
    "langchain-weaviate>=0.0.3",
    "markdown>=3.10",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "weaviate-client>=4.18.0",
    "pypdf>=6.2.0",
    "ipykernel>=7.1.0",
    "authlib>=1.6.5",
]
